refactor: renamed modal and sub-component to be generic
